Add favicon
I didn't really spend much time on this favicon, but it's better than
nothing.
I should probably design a better one in the future.

In zola-aurora we should probably also add fallback support for png
favicons, if some browsers don't support SVG icons.
This can easily be automated with the resize_image() [0] function and
its format argument, so that we don't have to add png artifacts to git.

[0] https://www.getzola.org/documentation/content/image-processing/